feat(pie-css): DSW-1166 scss mixins and functions for line height
