+++ .changeset/loud-parrots-exist.md
@@ -1,0 +1,5 @@
+---
+"@justeattakeaway/pie-css": minor
+---
+
+[Added] - SCSS mixins and functions for line-height
