--- packages/tools/pie-css/scss/functions/_fontSize.scss
@@ -1,20 +1,0 @@
-// ----------------------------------------------
-// Function for calculating the font size.
-// The `font-size` function allows you to compute the font size using a CSS variable.
-// The resulting value will be multiplied by 1px.
-// -
-// @param {String} $token - The CSS variable name (including the `--` prefix) to be used for the font size calculation.
-// -
-// @example
-// .my-class {
-//   font-size: font-size(--font-size-variable);
-// }
-// -
-// Outputs:
-// .my-class {
-//   font-size: calc(var(--font-size-variable) * 1px);
-// }
-// ----------------------------------------------
-@function font-size($token) {
-    @return calc(var(#{$token}) * 1px);
-}
