+++ packages/tools/pie-css/scss/functions/index.scss
@@ -1,1 +1,1 @@
-@forward './fontSize';
+@forward './unitModifiers';
