+++ packages/tools/pie-css/scss/mixins/index.scss
@@ -1,4 +1,4 @@
 @forward './buttonInteractiveStates';
-@forward './fontSize';
+@forward './unitModifiers';
 @forward './focus';
 @forward './visuallyHidden';
