+++ packages/tools/pie-css/test/scss/unit/mixins/unitModifiers.spec.ts
@@ -39,3 +39,35 @@ describe('mixins.font-size', () => {
     });
 });
 
+describe('mixins.line-height', () => {
+    it('should render the expected CSS content', async () => {
+        // Arrange
+        const scssToTest = `
+          @use 'mixins';
+
+          :root {
+            --line-height: 12;
+          }
+
+          .foo {
+            @include mixins.line-height(--line-height);
+          }
+        `;
+
+        const expectedCss = stripCSSWhitespace(`
+          :root {
+            --line-height: 12;
+          }
+
+          .foo {
+            line-height: calc(var(--line-height) * 1px);
+          }
+        `);
+
+        // Act
+        const css = stripCSSWhitespace(compileCss(scssToTest));
+
+        // Assert
+        expect(css).toBe(expectedCss);
+    });
+});
