+++ packages/tools/pie-css/test/scss/unit/validityTest.scss
@@ -6,11 +6,14 @@
 
 :root {
     --test-font-size: 12;
+    --test-line-height: 16;
     --bar-font-size: #{functions.font-size('--test-font-size')};
+    --bar-line-height: #{functions.line-height('--test-line-height')};
 }
 
 .foo {
     @include mixins.font-size(--font-size);
+    @include mixins.line-height(--line-height);
 
     :focus-visible {
         @include mixins.focus;
@@ -27,4 +30,5 @@
 
 .bar {
     font-size: var(--bar-font-size);
+    line-height: var(--bar-line-height);
 }
